Fixed segfault and added MonoBridge::Create()
The original example from DotNetConf2015 is wrong, you cannot close all the images prior
to unloading the appdomain, or else, the domain unloading thread will try to double free
the same image, which causes a segfault.

Also, inside the mono_domain_unload() code, it is already doing a GC, so removed redudant code.

Added Create() function which allows us to create an object and return it for use.